@@ -33,7 +33,7 @@ if t.TYPE_CHECKING:
 
     from _typeshed.wsgi import WSGIEnvironment
 
-    t_parse_result = t.Tuple[
+    t_parse_result = tuple[
         t.IO[bytes], MultiDict[str, str], MultiDict[str, FileStorage]
     ]
 
